fix: "cambio de longs a integer y correciones de enum en historial status"

+++ src/main/java/edu/dosw/dto/ReportDTO.java
@@ -2,10 +2,10 @@ package edu.dosw.dto;
 
 /** DTO for reporting reassignment statistics */
 public record ReportDTO(
-    long total,
-    long pending,
-    long approved,
-    long rejected,
-    long cancellations,
-    long swaps,
-    long joins) {}
+     Integer total,
+    Integer pending,
+    Integer approved,
+    Integer rejected,
+    Integer cancellations,
+    Integer swaps,
+     Integer joins) {}
